FIX: invalid implicit type conversion
I have figured out how to solve the problem of having an address in the
Data struct and then trying to use it as address payable. I just needed
to wrap payable() around the address in question. ToDo: do that for
another instance in the code where I found a different solution that is
going to be inconvenient down the line if I don't refactor the code.